+++ EngineClient/src/main/java/unknowndomain/engine/client/camera/FirstPersonController.java
@@ -1,5 +1,6 @@
 package unknowndomain.engine.client.camera;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 import unknowndomain.engine.client.display.Camera;
 
@@ -14,7 +15,7 @@ public class FirstPersonController extends CameraController {
     }
 
     @Override
-    public void update(Vector3f position, Vector3f rotation) {
+    public void update(Vector3d position, Vector3f rotation) {
         camera.getPosition().set(position);
 
         Vector3f front = new Vector3f((float) (Math.cos(Math.toRadians(this.pitch)) * Math.cos(Math.toRadians(this.yaw))), (float) Math.sin(Math.toRadians(this.pitch)), (float) (Math.cos(Math.toRadians(this.pitch)) * Math.sin(Math.toRadians(this.yaw)))).normalize();
